refactor: simplify GraphQL greeting logic and remove unused resolvers
