--- crates/rust-lambda-graphql/src/resolvers.rs
@@ -1,1 +1,0 @@
-pub mod greet;
